Android  Crystal Ball Update
With the crystal ball app i have done all the starter milestones, all the intermediate milestones, and all the advanced milestones up to the shake delay and the re-skining of the entire app. I have also added my own touch with the words appearing after the animation is done playing and fading away.